:sparkles: feat: Modal Service

+++ src/app/modules/maintenance/pages/clients-table-page/clients-table-page.component.ts
@@ -3,6 +3,7 @@ import { Client } from 'src/app/core/models/client';
 import { Page } from 'src/app/core/models/page';
 import { ClientService } from 'src/app/core/services/client.service';
 import { ToastService } from 'src/app/core/services/toast.service';
+import { ModalComponent } from 'src/app/shared/components/modal/modal.component';
 
 @Component({
   selector: 'app-clients-table-page',
@@ -13,6 +14,7 @@ export class ClientsTablePageComponent implements OnInit {
   clientPage: Page<Client> = {} as Page<Client>;
   page = 1;
   nameFilter: string = '';
+  selectedClient!: Client;
 
   constructor(
     private clientService: ClientService,
@@ -42,19 +44,24 @@ export class ClientsTablePageComponent implements OnInit {
     this.loadClients();
   }
 
-  delete(client: Client): void {
-    this.clientService.delete(client).subscribe({
-      next: () => {
-        this.loadClients();
-        this.toastService.show('Cliente removido com sucesso!', {
-          classname: 'bg-success text-light',
+  delete(client: Client, modalConfirm: ModalComponent): void {
+    this.selectedClient = client;
+    modalConfirm.open().then((confirm) => {
+      if (confirm) {
+        this.clientService.delete(client).subscribe({
+          next: () => {
+            this.loadClients();
+            this.toastService.show('Cliente removido com sucesso!', {
+              classname: 'bg-success text-light',
+            });
+          },
+          error: () => {
+            this.toastService.show('Erro ao remover o cliente!', {
+              classname: 'bg-danger text-light',
+            });
+          },
         });
-      },
-      error: () => {
-        this.toastService.show('Erro ao remover o cliente!', {
-          classname: 'bg-danger text-light',
-        });
-      },
+      }
     });
   }
 }
